Move toml/yaml serialization to dvc.utils.serialize
