+++ dvc/utils/serialize/__init__.py
@@ -1,0 +1,2 @@
+from ._toml import *  # noqa, pylint: disable=wildcard-import
+from ._yaml import *  # noqa, pylint: disable=wildcard-import
