+++ tests/func/test_add.py
@@ -27,7 +27,7 @@ from dvc.system import System
 from dvc.tree.local import LocalTree
 from dvc.utils import LARGE_DIR_SIZE, file_md5, relpath
 from dvc.utils.fs import path_isin
-from dvc.utils.yaml import load_yaml
+from dvc.utils.serialize import load_yaml
 from tests.basic_env import TestDvc
 from tests.utils import get_gitignore_content
